Add Metadata generator (#3)
Add a new script that generates metadata for MCAP
files. The script reads MCAP files, extracts metadata such as duration,
topic message counts, and message types, and compiles this information
into a JSON file. It also includes functionalities to get the file size
and compute file hash. A command-line interface is provided for ease of
use.